LayoutTests/media/modern-media-controls/macos-fullscreen-media-controls/macos-fullscreen-media-controls-buttons-containers-styles.html

<script src="../../../resources/js-test-pre.js"></script>
<script src="../resources/media-controls-loader.js" type="text/javascript"></script>
<body>
<div id="container"></div>
<style type="text/css" media="screen">

#container {
    position: absolute;
    top: 0;
    left: 0;
    width: 800px;
    height: 600px;
}

</style>
<script type="text/javascript">

description("Testing the <code>MacOSFullscreenMediaControls</code> computed styles.");

window.jsTestIsAsync = true;

const mediaControls = new MacOSFullscreenMediaControls;
const centerContainer = mediaControls.controlsBar.children[1];
const rightContainer = mediaControls.controlsBar.children[2];
const buttons = centerContainer.buttons.concat(rightContainer.buttons);

let centerContainerBounds, rightContainerBounds;
scheduler.frameDidFire = function()
{
    if (buttons.some(button => button.width == 0))
        return;

    // All buttons have been loaded, we wait for the next tick to ensure
    // that their CSS properties have also been committed.
    window.requestAnimationFrame(() => {
        document.getElementById("container").appendChild(mediaControls.element);

        centerContainerBounds = centerContainer.element.getBoundingClientRect();

        shouldBe("centerContainerBounds.left", "311");
        shouldBe("centerContainerBounds.top", "500");
        shouldBe("centerContainerBounds.width", "178");
        shouldBe("centerContainerBounds.height", "44");

        debug("");
        rightContainerBounds = rightContainer.element.getBoundingClientRect();
        shouldBeCloseTo("rightContainerBounds.left", 496, 0.5);
        shouldBe("rightContainerBounds.top", "500");
        shouldBeCloseTo("rightContainerBounds.width", 138, 0.5);
        shouldBe("rightContainerBounds.height", "44");

        debug("");
        mediaControls.element.remove();
        finishJSTest();
    });

    scheduler.frameDidFire = null;
};

</script>
<script src="../../../resources/js-test-post.js"></script>
</body>
